[feat/TK-128]: ShowMyTotalPaymentResponse 생성 로직 구현
- TravelMemberSettlementHistory 엔티티의 ownPaymentCost, actualBurdenCost의 타입을 double -> int로 변경 (원화)
- SettlementErrorCode에 MEMBER_SETTLEMENT_HISTORY_NOT_FOUND 에러 추가

+++ src/main/java/withbeetravel/domain/TravelMemberSettlementHistory.java
@@ -26,10 +26,10 @@ public class TravelMemberSettlementHistory {
     private TravelMember travelMember;
 
     @Column(name = "own_payment_cost", nullable = false)
-    private double ownPaymentCost;
+    private int ownPaymentCost;
 
     @Column(name = "actual_burden_cost", nullable = false)
-    private double actualBurdenCost;
+    private int actualBurdenCost;
 
     @Column(name = "is_agreed", nullable = false)
     private boolean isAgreed;
@@ -38,8 +38,8 @@ public class TravelMemberSettlementHistory {
     public TravelMemberSettlementHistory(Long id,
                                          SettlementRequest settlementRequest,
                                          TravelMember travelMember,
-                                         double ownPaymentCost,
-                                         double actualBurdenCost,
+                                         int ownPaymentCost,
+                                         int actualBurdenCost,
                                          boolean isAgreed) {
         this.id = id;
         this.settlementRequest = settlementRequest;
